feat(cli): adding cli with a convention for cmds
