--- cmd/examples/main.go
@@ -1,9 +1,0 @@
-// Package main provides the entry point for the workflow engine application.
-// It demonstrates the usage of the workflow engine with example workflows.
-package main
-
-import "github.com/open-source-cloud/fuse/internal/app"
-
-func main() {
-	app.RunServerCli()
-}
